==> code_revision/01. Stock crawling with yf.ipynb <==
import pandas as pd
import yfinance as yf

NASDAQ = pd.read_csv("../data/Stock_info/NASDAQ.csv")
NASDAQ.sort_values(by=["Market Cap"], ascending=False, inplace=True)
NASDAQ_symbols = NASDAQ[(NASDAQ["IPO Year"]<2009).values & (NASDAQ["IPO Year"].notna()).values & (NASDAQ["Name"].str.contains("Common Stock").values)][0:100]["Symbol"].values

NASDAQ_stock = yf.download(NASDAQ_symbols.tolist(), start="2009-01-02", end="2023-11-24")

NASDAQ_stock

NASDAQ_stock.reset_index().to_csv("../data/Stock_data/NASDAQ.csv", index=False)

KOSPI = pd.read_csv("../data/Stock_info/KOSPI.csv", encoding='euc-kr')
KOSPI['종목코드'] = KOSPI['종목코드'].astype(str).str.zfill(6) + ".KS"
KOSPI_symbols = KOSPI.종목코드.values[:100]

KOSPI_stock = yf.download(KOSPI_symbols.tolist(), start="2009-01-02", end="2023-11-24")

KOSPI_stock

KOSPI_stock.reset_index().to_csv("../data/Stock_data/KOSPI.csv", index=False)
